+++ src/main/java/eu/europa/ec/dgc/businessrule/restapi/controller/BusinessRuleController.java
@@ -22,22 +22,17 @@ package eu.europa.ec.dgc.businessrule.restapi.controller;
 
 import eu.europa.ec.dgc.businessrule.entity.BusinessRuleEntity;
 import eu.europa.ec.dgc.businessrule.exception.DgcaBusinessRulesResponseException;
-import eu.europa.ec.dgc.businessrule.model.BusinessRuleItem;
 import eu.europa.ec.dgc.businessrule.restapi.dto.BusinessRuleListItemDto;
 import eu.europa.ec.dgc.businessrule.restapi.dto.ProblemReportDto;
 import eu.europa.ec.dgc.businessrule.service.BusinessRuleService;
-import eu.europa.ec.dgc.businessrule.utils.BusinessRulesUtils;
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.headers.Header;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import java.security.NoSuchAlgorithmException;
 import java.util.List;
 import java.util.Locale;
 import javax.validation.Valid;
@@ -48,8 +43,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -66,8 +59,6 @@ public class BusinessRuleController {
     private static final String X_SIGNATURE_HEADER = "X-SIGNATURE";
 
     private final BusinessRuleService businessRuleService;
-
-    private final BusinessRulesUtils businessRulesUtils;
 
     /**
      * Http Method for getting the business rules list.
@@ -240,47 +231,10 @@ public class BusinessRuleController {
         return ResponseEntity.ok(rule.getRawData());
     }
 
-    /**
-     * Http Method for uploading sample data.
-     */
-    @PostMapping(path = "", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    @Hidden
-    public ResponseEntity<String> createRule(
-        @RequestHeader(value = "X_COUNTRY") String country,
-        @RequestHeader(value = "X_ID") String id,
-        @RequestHeader(value = "X_VER") String version,
-        @RequestBody String ruleData) {
-
-        String hash;
-        validateCountryParameter(country);
-
-        if (id == null || id.isBlank()) {
-            throw new DgcaBusinessRulesResponseException(HttpStatus.BAD_REQUEST, "0x003", "Possible reasons: "
-                + "The id of the rule is not set.", id,"");
-        }
-
-        try {
-            hash = businessRulesUtils.calculateHash(ruleData);
-        } catch (NoSuchAlgorithmException e) {
-            log.error("Calculation of hash failed:", e);
-            throw new DgcaBusinessRulesResponseException(HttpStatus.INTERNAL_SERVER_ERROR, "0x500",
-                "Internal Server Error","","");
-        }
-        BusinessRuleItem bri = new BusinessRuleItem();
-        bri.setHash(hash);
-        bri.setIdentifier(id);
-        bri. setCountry(country);
-        bri.setVersion(version);
-        bri.setRawData(ruleData);
-        businessRuleService.saveBusinessRule(bri);
-        return ResponseEntity.ok("Upload: OK");
-
-    }
-
     private void validateCountryParameter(String country) throws DgcaBusinessRulesResponseException {
         if (!country.matches("^[a-zA-Z]{2}$")) {
             throw new DgcaBusinessRulesResponseException(HttpStatus.BAD_REQUEST, "0x004", "Possible reasons: "
-                + "The Country Code has a wrong format. Should be 2 char format.", country,"");
+                    + "The Country Code has a wrong format. Should be 2 char format.", country,"");
         }
     }
